macro-bot/requirements.txt
yfinance
fredapi
pandas
schedule
google-genai
